requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
pytest==6.2.4
pytest-flask==1.2.0
psycopg2-binary==2.9.1  # Se você estiver usando PostgreSQL
pymysql==1.0.2          # Se você estiver usando MySQL
